initial vehicle model and sheet
