--- scripts/actor/model/base.js
@@ -1,15 +1,0 @@
-
-export class BaseWNGActorModel extends foundry.abstract.DataModel {
-
-
-    computeBase() 
-    {
-
-    }
-
-    computeDerived(items, autoCalc) 
-    {
-        
-    }
-
-}
