+++ scripts/model/actor/components/attributes.js
