+++ scripts/model/actor/components/base.js
@@ -1,0 +1,19 @@
+
+export class BaseWnGActorModel extends foundry.abstract.DataModel {
+
+    static defineSchema() 
+    {
+        return {};
+    }
+
+    computeBase() 
+    {
+
+    }
+
+    computeDerived(items, autoCalc) 
+    {
+        
+    }
+
+}
