+++ static/template/actor/threat.html
@@ -1,78 +1,76 @@
 <form class="{{cssClass}}" autocomplete="off">
-  <div class="threat">
-    <div class="sheet-header">
-      <div class="avatar">
-        <img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}" />
+  <div class="sheet-header">
+    <div class="avatar">
+      <img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}" />
+    </div>
+    <div class="details">
+      <div class="row">
+        <div class="name">
+          <label>{{localize "BIO.NAME"}}</label>
+          <input name="name" type="text" value="{{actor.name}}" />
+        </div>
+        <div class="species">
+          <label class="item-label" data-type="species">{{localize "BIO.SPECIES"}}</label>
+          <input disabled type="text" value="{{actor.species.name}}" />
+        </div>
       </div>
-      <div class="bio">
-        <div class="wrapper-row">
-          <div class="name">
-            <label>{{localize "BIO.NAME"}}</label>
-            <input name="name" type="text" value="{{actor.name}}" />
-          </div>
-          <div class="species">
-            <label class="item-label" data-type="species">{{localize "BIO.SPECIES"}}</label>
-            <input  disabled type="text" value="{{actor.species.name}}" />
-          </div>
+      <div class="row">
+        <div class="faction">
+          <label class="item-label" data-type="faction">{{localize "BIO.FACTION"}}</label>
+          <input disabled type="text" value="{{actor.faction.name}}" />
         </div>
-        <div class="wrapper-row">
-          <div class="faction">
-            <label class="item-label" data-type="faction">{{localize "BIO.FACTION"}}</label>
-            <input  disabled type="text" value="{{actor.faction.name}}" />
-          </div>
-          <div class="archetype">
-            <label class="item-label" data-type="archetype">{{localize "BIO.ARCHETYPE"}}</label>
-            <input disabled type="text" value="{{actor.archetype.name}}" />
-          </div>
+        <div class="archetype">
+          <label class="item-label" data-type="archetype">{{localize "BIO.ARCHETYPE"}}</label>
+          <input disabled type="text" value="{{actor.archetype.name}}" />
         </div>
-        <div class="wrapper-row">
-          <div class="threat">
-            <label>{{localize "BIO.THREAT"}}</label>
-            <input name="system.bio.threat.first" type="text" value="{{system.bio.threat.first}}" />
-            <input name="system.bio.threat.second" type="text" value="{{system.bio.threat.second}}" />
-            <input name="system.bio.threat.third" type="text" value="{{system.bio.threat.third}}" />
-            <input name="system.bio.threat.fourth" type="text" value="{{system.bio.threat.fourth}}" />
-          </div>
-          <div class="ruin">
-            <label>{{localize "RESOURCE.RUIN"}}</label>
-            <input name="system.resources.ruin" type="number" value="{{system.resources.ruin}}" data-dtype="Number" />
-          </div>
-          <div class="mob">
-            <label>{{localize "THREAT.MOB"}}</label>
-            <input name="system.mob" type="number" value="{{system.mob}}" data-dtype="Number" />
-          </div>
+      </div>
+      <div class="row">
+        <div class="threat">
+          <label>{{localize "BIO.THREAT"}}</label>
+          <input name="system.bio.threat.first" type="text" value="{{system.bio.threat.first}}" />
+          <input name="system.bio.threat.second" type="text" value="{{system.bio.threat.second}}" />
+          <input name="system.bio.threat.third" type="text" value="{{system.bio.threat.third}}" />
+          <input name="system.bio.threat.fourth" type="text" value="{{system.bio.threat.fourth}}" />
+        </div>
+        <div class="ruin">
+          <label>{{localize "RESOURCE.RUIN"}}</label>
+          <input name="system.resources.ruin" type="number" value="{{system.resources.ruin}}" data-dtype="Number" />
+        </div>
+        <div class="mob">
+          <label>{{localize "THREAT.MOB"}}</label>
+          <input name="system.mob" type="number" value="{{system.mob}}" data-dtype="Number" />
         </div>
       </div>
     </div>
+  </div>
 
-    <div class="sheet-tabs tabs" data-group="primary">
-      <b class="item" data-tab="stats">{{localize "TAB.STATS"}}</b>
-      <b class="item" data-tab="combat">{{localize "TAB.COMBAT"}}</b>
-      <b class="item" data-tab="talents">{{localize "TAB.TALENTS"}}</b>
-      <b class="item" data-tab="gear">{{localize "TAB.GEAR"}}</b>
-      <b class="item" data-tab="effects">{{localize "TAB.EFFECTS"}}</b>
-      <b class="item" data-tab="notes">{{localize "TAB.NOTES"}}</b>
+  <div class="sheet-tabs tabs" data-group="primary">
+    <b class="item" data-tab="stats">{{localize "TAB.STATS"}}</b>
+    <b class="item" data-tab="combat">{{localize "TAB.COMBAT"}}</b>
+    <b class="item" data-tab="talents">{{localize "TAB.TALENTS"}}</b>
+    <b class="item" data-tab="gear">{{localize "TAB.GEAR"}}</b>
+    <b class="item" data-tab="effects">{{localize "TAB.EFFECTS"}}</b>
+    <b class="item" data-tab="notes">{{localize "TAB.NOTES"}}</b>
+  </div>
+
+  <div class="sheet-body">
+    <div class="tab" data-group="primary" data-tab="stats">
+      {{> systems/wrath-and-glory/template/actor/tab/stats.html}}
     </div>
-
-    <div class="sheet-body">
-      <div class="tab" data-group="primary" data-tab="stats">
-        {{> systems/wrath-and-glory/template/actor/tab/stats.html}}
-      </div>
-      <div class="tab" data-group="primary" data-tab="talents">
-        {{> systems/wrath-and-glory/template/actor/tab/talents.html}}
-      </div>
-      <div class="tab" data-group="primary" data-tab="combat">
-        {{> systems/wrath-and-glory/template/actor/tab/combat.html}}
-      </div>
-      <div class="tab" data-group="primary" data-tab="gear">
-        {{> systems/wrath-and-glory/template/actor/tab/gear.html}}
-      </div>
-      <div class="tab" data-group="primary" data-tab="effects">
-        {{> systems/wrath-and-glory/template/actor/tab/effects.html}}
-      </div>
-      <div class="tab" data-group="primary" data-tab="notes">
-        {{> systems/wrath-and-glory/template/actor/tab/notes-threat.html}}
-      </div>
+    <div class="tab" data-group="primary" data-tab="talents">
+      {{> systems/wrath-and-glory/template/actor/tab/talents.html}}
+    </div>
+    <div class="tab" data-group="primary" data-tab="combat">
+      {{> systems/wrath-and-glory/template/actor/tab/combat.html}}
+    </div>
+    <div class="tab" data-group="primary" data-tab="gear">
+      {{> systems/wrath-and-glory/template/actor/tab/gear.html}}
+    </div>
+    <div class="tab" data-group="primary" data-tab="effects">
+      {{> systems/wrath-and-glory/template/actor/tab/effects.html}}
+    </div>
+    <div class="tab" data-group="primary" data-tab="notes">
+      {{> systems/wrath-and-glory/template/actor/tab/notes-threat.html}}
     </div>
   </div>
 </form>
