+++ static/template/actor/vehicle.html
@@ -1,0 +1,68 @@
+<form class="{{cssClass}}" autocomplete="off">
+  <div class="sheet-header">
+    <div class="avatar">
+      <img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}" />
+    </div>
+    <div class="details">
+      <div class="row">
+        <div class="name">
+          <label>{{localize "Name"}}</label>
+          <input name="name" type="text" value="{{actor.name}}" />
+        </div>
+        <div class="size">
+          <label class="item-label">{{localize "HEADER.SIZE"}}</label>
+          <select name="system.combat.size">
+            {{#select system.combat.size}}
+              {{#each (config "size")}}
+                <option value="{{@key}}">{{this}}</option>
+              {{/each}}
+            {{/select}}
+          </select>
+        </div>
+      </div>
+      <div class="row">
+        <div class="speed">
+          <label>{{localize "HEADER.SPEED"}}</label>
+          <input type="number" value="{{system.combat.speed}}" name="system.combat.speed" />
+        </div>
+        <div class="resilience">
+          <label>{{localize "HEADER.RESILIENCE"}}</label>
+          <input type="number" value="{{system.combat.resilience.total}}" name="system.combat.resilience.total" />
+        </div>
+      </div>
+      <div class="row">
+        <div class="defence">
+          <label>{{localize "HEADER.DEFENCE"}}</label>
+          <input type="number" value="{{system.defence.total}}" />
+        </div>
+        <div class="wounds">
+          <label>{{localize "HEADER.WOUNDS"}}</label>
+          <input type="number" value="{{system.combat.wounds.value}}" />
+          <input type="number" value="{{system.combat.wounds.max}}" />
+        </div>
+      </div>
+    </div>
+  </div>
+
+  <div class="sheet-tabs tabs" data-group="primary">
+    <b class="item" data-tab="main">{{localize "TAB.MAIN"}}</b>
+    <b class="item" data-tab="gear">{{localize "TAB.GEAR"}}</b>
+    <b class="item" data-tab="effects">{{localize "TAB.EFFECTS"}}</b>
+    <b class="item" data-tab="notes">{{localize "TAB.NOTES"}}</b>
+  </div>
+
+  <div class="sheet-body">
+    <div class="tab" data-group="primary" data-tab="stats">
+      {{> systems/wrath-and-glory/template/actor/tab/vehicle/main.html}}
+    </div>
+    <div class="tab" data-group="primary" data-tab="gear">
+      {{> systems/wrath-and-glory/template/actor/tab/vehicle/gear.html}}
+    </div>
+    <div class="tab" data-group="primary" data-tab="effects">
+      {{> systems/wrath-and-glory/template/actor/tab/effects.html}}
+    </div>
+    <div class="tab" data-group="primary" data-tab="notes">
+      {{> systems/wrath-and-glory/template/actor/tab/notes-threat.html}}
+    </div>
+  </div>
+</form>
