+++ static/template/apps/item-dialog.html
@@ -1,0 +1,10 @@
+  <div class="directory">
+    <ol class="directory-list">
+      {{#each items as |item i|}}
+      <li class="directory-item document flexrow" data-index="{{i}}" data-id="{{item.id}}">
+        <img class="thumbnail" title="{{item.name}}" src="{{item.img}}">
+        <h4 class="document-name"><a>{{item.name}}</a></h4>
+      </li>
+      {{/each}}
+    </ol>
+  </div>
