+++ style/sheet/_threat.scss
@@ -1,135 +1,21 @@
-.wrath-and-glory.sheet.actor .threat {
-  width: 100%;
-  height: 100%;
+.wrath-and-glory.sheet.actor.threat {
+  .details {
+
+    .threat input {
+      flex-basis: 20%;
+    }
+
+    .mob,
+    .ruin {
+      display: flex;
+      flex-direction: row;
+      flex-basis: calc(20% - 1px);
+      margin-left: 1px;
+
+      label {
+        min-width: 50px;
+        max-width: 50px;
+      }
+    }
+  }
 }
-
-/* HEADER */
-
-.wrath-and-glory.sheet.actor .threat .sheet-header {
-  display: grid;
-  grid-template-columns: 100px auto;
-  grid-template-rows: 100px;
-  column-gap: 8px;
-  row-gap: 0;
-  width: 100%;
-  height: 100px;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .avatar {
-  grid-column: 1 / 1;
-  grid-row: 1 / 1;
-  border: 8px solid transparent;
-  border-image: var(--image-border) 8 8 8 8 repeat repeat;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .avatar img {
-  border: 0 none;
-  height: 100%;
-  width: 100%;
-  object-fit: contain;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  flex-grow: 1;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .wrapper-row {
-  display: flex;
-  flex-direction: row;
-  flex-basis: 100%;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio label {
-  min-width: 100px;
-  max-width: 100px;
-  height: 31px;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio input {
-  flex-grow: 1;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .name,
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .faction {
-  display: flex;
-  flex-direction: row;
-  flex-basis: calc(50% - 1px);
-  margin-right: 1px;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .species,
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .archetype {
-  display: flex;
-  flex-direction: row;
-  flex-basis: calc(50% - 1px);
-  margin-left: 1px;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .threat {
-  display: flex;
-  flex-direction: row;
-  flex-basis: calc(80% - 1px);
-  margin-right: 1px;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .threat input {
-  flex-basis: 20%;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .mob,
-  .wrath-and-glory.sheet.actor .threat .sheet-header .bio .ruin {
-  display: flex;
-  flex-direction: row;
-  flex-basis: calc(20% - 1px);
-  margin-left: 1px;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .ruin label {
-  min-width: 50px;
-  max-width: 50px;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-header .bio .mob label {
-  min-width: 50px;
-  max-width: 50px;
-}
-
-
-/* TABS */
-
-.wrath-and-glory.sheet.actor .threat .sheet-tabs {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  flex-basis: 100%;
-  font-size: 18px;
-  font-family: var(--font-text);
-  font-weight: bold;
-  margin: 8px 0;
-  border: 8px solid transparent;
-  border-image: var(--image-border) 8 8 8 8 repeat repeat;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-tabs .item {
-  color: var(--color-white);
-  background-color: var(--color-black);
-  cursor: pointer;
-  padding: 5px 0;
-  flex: 1;
-}
-
-.wrath-and-glory.sheet.actor .threat .sheet-tabs .item:hover,
-.wrath-and-glory.sheet.actor .threat .sheet-tabs .item.active  {
-  background-color: var(--color-red);
-  text-shadow: none;
-}
-
-/* BODY */
-
-/* .wrath-and-glory.sheet.actor .threat .sheet-body {
-  width: 100%;
-  height: 588px;
-} */
