+++ style/sheet/_vehicle.scss
@@ -1,0 +1,9 @@
+.wrath-and-glory.sheet.actor.vehicle {
+    .sheet-header .details .row {
+      .size {
+        select {
+          width: 100%;
+        }
+      }
+    }
+  }
