+++ style/wrath-and-glory.scss
@@ -3,8 +3,7 @@
 @import "./common/sheet";
 @import "./common/variable";
 @import "./dialog/dialog";
-@import "./sheet/agent";
-@import "./sheet/threat";
+@import "./sheet/actor";
 @import "./sheet/tab/combat";
 @import "./sheet/tab/gear";
 @import "./sheet/tab/notes";
